+++ directive/index.js
@@ -1,19 +1,19 @@
 'use strict';
-var path = require('path');
 var util = require('util');
 var ScriptBase = require('../script-base.js');
-var angularUtils = require('../util.js');
 
 
-module.exports = Generator;
-
-function Generator() {
+var Generator = module.exports = function Generator() {
   ScriptBase.apply(this, arguments);
-}
+};
 
 util.inherits(Generator, ScriptBase);
 
 Generator.prototype.createDirectiveFiles = function createDirectiveFiles() {
-  this.appTemplate('directive');
-  this.testTemplate('spec/directive');
+  this.generateSourceAndTest(
+    'directive',
+    'spec/directive',
+    'directives',
+    this.options['skip-add'] || false
+  );
 };