+++ route/index.js
@@ -5,44 +5,35 @@ var ScriptBase = require('../script-base.js');
 var angularUtils = require('../util.js');
 
 
-module.exports = Generator;
-
-function Generator() {
+var Generator = module.exports = function Generator() {
   ScriptBase.apply(this, arguments);
   this.hookFor('angular:controller');
   this.hookFor('angular:view');
-}
+};
+
 util.inherits(Generator, ScriptBase);
 
 Generator.prototype.rewriteAppJs = function () {
-  var htmlTemplatePath = this.name + '.html',
-    splicable,
-    filePath = path.join(this.env.options.appPath, 'scripts/app.');
-
-  if (htmlTemplatePath.indexOf('/') === -1) {
-    htmlTemplatePath = 'views/' + htmlTemplatePath;
-  }
-
-  if (this.env.options.coffee) {
-    splicable = [
-      '.when \'/' + this.name + '\',',
-      '  templateUrl: \'' + htmlTemplatePath + '\',',
-      '  controller: \'' + this._.classify(this.name) + 'Ctrl\''
-    ];
-    filePath += 'coffee';
-  } else {
-    splicable = [
-      '.when(\'/' + this.name + '\', {',
-      '  templateUrl: \'' + htmlTemplatePath + '\',',
-      '  controller: \'' + this._.classify(this.name) + 'Ctrl\'',
-      '})'
-    ];
-    filePath += 'js';
-  }
-
-  angularUtils.rewriteFile({
-    file: filePath,
+  var coffee = this.env.options.coffee;
+  var config = {
+    file: path.join(
+      this.env.options.appPath,
+      'scripts/app.' + (coffee ? 'coffee' : 'js')
+    ),
     needle: '.otherwise',
-    splicable: splicable
-  });
+    splicable: [
+      "  templateUrl: 'views/" + this.name + ".html'" + (coffee ? "" : "," ),
+      "  controller: '" + this.classedName + "Ctrl'"
+    ]
+  };
+
+  if (coffee) {
+    config.splicable.unshift(".when '/" + this.name + "',");
+  }
+  else {
+    config.splicable.unshift(".when('/" + this.name + "', {");
+    config.splicable.push("})");
+  }
+
+  angularUtils.rewriteFile(config);
 };