+++ view/index.js
@@ -4,9 +4,7 @@ var util = require('util');
 var yeoman = require('yeoman-generator');
 
 
-module.exports = Generator;
-
-function Generator() {
+var Generator = module.exports = function Generator() {
   yeoman.generators.NamedBase.apply(this, arguments);
   this.sourceRoot(path.join(__dirname, '../templates'));
 
@@ -16,14 +14,10 @@ function Generator() {
     } catch (e) {}
     this.env.options.appPath = this.env.options.appPath || 'app';
   }
-}
+};
 
 util.inherits(Generator, yeoman.generators.NamedBase);
 
 Generator.prototype.createViewFiles = function createViewFiles() {
-  var targetPath = this.name;
-  if (this.name.indexOf('/') === -1) {
-    targetPath = 'views/' + targetPath;
-  }
-  this.template('common/view.html', path.join(this.env.options.appPath, targetPath + '.html'));
+  this.template('common/view.html', path.join(this.env.options.appPath, 'views', this.name + '.html'));
 };